requirements.txt
requests
websocket-client
PyNaCl
python-dotenv
base58
